## What this PR does / why we need it:

Add support for Bitbucket Cloud in the Terramate Cloud sync features.

## Which issue(s) this PR fixes:
none

## Special notes for your reviewer:

## Does this PR introduce a user-facing change?
```
yes, adds a Terramate Cloud feature.
```

+++ cmd/terramate/cli/gitlab/mr.go
@@ -35,7 +35,7 @@ type (
 		// Token is the Gitlab token (usually provided by the GITLAB_TOKEN environment variable.
 		Token string
 
-		ProjectID int
+		ProjectID int64
 		Group     string
 		Project   string
 	}
